+++ pulse_lib/segments/segment_container.py
@@ -167,6 +167,10 @@ class segment_container():
         add_reference_channels(new, self._virtual_gate_matrices, self._IQ_channel_objs)
         return new
 
+    def add(self, other, time=None):
+        for name in self.channels.keys():
+            self[name].add(other[name], time)
+
     @property
     def software_markers(self):
         return self._software_markers