+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to Pulselib will be documented in this file.
 
+## \[1.6.8] - 2023-02-02
+
+- Fixed setpoints of measurment parameter for time traces > 2 seconds.
+- Fixed simultaneous pulses with indexing of segments on looping parameter.
+- Calculate timeout for get_acquisition_data from sequence.
+
+## \[1.6.7] - 2023-02-01
+
+- Added continuous mode for Keysight AWG.
+
 ## \[1.6.6] - 2023-01-30
 
 - Fixed errors when slicing segments after adding dimension with update_dim.